+++ BTDB/IOC/CReg.cs
@@ -8,8 +8,10 @@ class CReg
     internal Func<IContainer, ICreateFactoryCtx, Func<IContainer, IResolvingCtx?, object>> Factory;
     internal Lifetime Lifetime;
     internal uint SingletonId;
+    internal uint ScopedId;
     internal StructList<CReg> Multi;
     internal Func<IContainer, IResolvingCtx?, object>? SingletonFactoryCache;
+    internal Func<IContainer, IResolvingCtx?, object>? ScopedFactoryCache;
 
     public void Add(CReg registration, bool preserveExistingDefaults)
     {
@@ -18,6 +20,7 @@ class CReg
             Factory = registration.Factory;
             Lifetime = registration.Lifetime;
             SingletonId = registration.SingletonId;
+            ScopedId = registration.ScopedId;
         }
 
         Multi.Add(registration);