+++ BTDB/IOC/IContainer.cs
@@ -3,8 +3,9 @@ using System.Collections.Concurrent;
 
 namespace BTDB.IOC;
 
-public interface IContainer
+public interface IContainer : IAsyncDisposable
 {
+    IContainer CreateScope();
     object Resolve(Type type);
     object ResolveNamed(string name, Type type);
     object ResolveKeyed(object key, Type type);