+++ BTDB/IOC/ILiveScopeTrait.cs
@@ -3,4 +3,5 @@ namespace BTDB.IOC;
 public interface ILiveScopeTrait
 {
     void SingleInstance();
+    void Scoped();
 }