+++ BTDB/IOC/SingleRegistration.cs
@@ -20,6 +20,11 @@ class SingleRegistration : RegistrationBaseImpl<IAsLiveScopeTrait>, IContanerReg
         _lifetime = Lifetime.Singleton;
     }
 
+    public void Scoped()
+    {
+        _lifetime = Lifetime.Scoped;
+    }
+
     public Lifetime Lifetime => _lifetime;
 
     public SingleRegistration(Type implementationType, bool withFallback = false)
@@ -73,7 +78,8 @@ class SingleRegistration : RegistrationBaseImpl<IAsLiveScopeTrait>, IContanerReg
         context.AddCReg(GetAsTypesFor(_implementationType), PreserveExistingDefaults, UniqueRegistration,
             new()
             {
-                Factory = factory, Lifetime = Lifetime, SingletonId = Lifetime == Lifetime.Singleton ? uint.MaxValue : 0
+                Factory = factory, Lifetime = Lifetime, SingletonId = Lifetime == Lifetime.Singleton ? uint.MaxValue : 0,
+                ScopedId = Lifetime == Lifetime.Scoped ? uint.MaxValue : 0
             });
     }